feat: init a ce client
